Final commit of Week 1. Added classes for vertices with different kinds of vertex data, quads, and base classes doing the API-calls to set up OpenGL.

+++ OpenGLexperiment/glewInit.cpp
@@ -2,7 +2,7 @@
 
 GLuint InitGlew (void) {
 	if(!glewInit()){
-		return 1001;
+		return (GLuint)1001;
 	}
-	return 0;
+	return (GLuint)0;
 }
